Return NULL, rather than assert in Sparse Memory alloc
In an unlikely case SparseVirtualMemory is fragmented or configured
inappropriately small, proceed with a failed allocation by returning
NULL rather than asserting it's not possible.

This will help with a caller to be able to handle the situation as
OutOfMemory exception (as any other allocation failure in the main
heap).

Signed-off-by: Aleksandar Micic <Aleksandar_Micic@ca.ibm.com>